requirements.txt
#kafka-python
#pyspark
streamlit
findspark
setuptools
six
kafka-python-ng